+++ CHANGELOG.md
@@ -92,6 +92,7 @@ CHANGELOG
 #### Outputs
 - `intelmq.bots.outputs.postgresql`: Recommend psycopg2-binary package.
 - `intelmq.bots.outputs.amqptopic.output`: Shutdown: Close connection only if connection exists.
+- `intelmq.bots.outputs.amqptopic`: Add support for pika > 1, the way the (Non-)Acknowledgments are provided has been changed.
 
 ### Documentation
 - Add certbund-contact to the ecosystem document.